+++ crates/ryft/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 - Added `ryft-ndarray` and `ryft-xla`.
 - Updated the `ryft-core`, `ryft-mlir`, `ryft-pjrt`, and `ryft-xla-sys` dependencies.
+- Added propagation for the `mps` feature of `ryft-xla-sys`.
 
 ## [0.0.2] - 2026-03-02